# openapi-specification.yaml
openapi: 3.1.0
info:
  title: ProcGrid B2B E-commerce Platform API
  description: |
    Production-grade B2B e-commerce platform for producer-buyer agricultural product trading.
    
    ## Authentication
    All protected endpoints require JWT Bearer token obtained from Keycloak authentication server.
    
    ## Rate Limiting
    - Public endpoints: 100 requests/minute
    - Authenticated endpoints: 1000 requests/minute
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details for HTTP APIs standard.
  version: 1.0.0
  contact:
    name: ProcGrid API Support
    email: api-support@procgrid.com
    url: https://procgrid.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://procgrid.com/terms
servers:
  - url: https://api.procgrid.com/v1
    description: Production server
  - url: https://staging-api.procgrid.com/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password via Keycloak
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              producer_login:
                summary: Producer login
                value:
                  email: "farmer@example.com"
                  password: "securePassword123"
                  rememberMe: false
              buyer_login:
                summary: Buyer login
                value:
                  email: "buyer@agrimart.com"
                  password: "buyerPass456"
                  rememberMe: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  value:
                    success: true
                    message: "Login successful"
                    data:
                      accessToken: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refreshToken: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                      tokenType: "Bearer"
                      expiresIn: 3600
                      userId: "12345"
                      email: "farmer@example.com"
                      firstName: "John"
                      lastName: "Farmer"
                      roles: ["PRODUCER"]
                      issuedAt: "2025-11-15T10:30:00Z"
                    timestamp: "2025-11-15T10:30:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout user and invalidate session
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
          example: "12345"
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Keycloak authentication
...
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        errorCode:
          type: string
          description: Error code for failed requests
          nullable: true
        data:
          description: Response data
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - errorCode
        - timestamp
        - path
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error description
        errorCode:
          type: string
          description: Specific error code
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint path

    # ============ AUTHENTICATION SCHEMAS ============
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: "farmer@example.com"
        password:
          type: string
          minLength: 6
          maxLength: 128
          example: "securePassword123"
        rememberMe:
          type: boolean
          default: false
...
  responses:
...
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            errorCode: "AUTH_003"
            timestamp: "2025-11-15T10:30:00Z"
            path: "/api/v1/orders"
...
    ValidationError:
      description: Validation error - Invalid input data
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  data:
                    type: object
                    additionalProperties:
                      type: string
          example:
            success: false
            message: "Validation failed"
            errorCode: "VAL_001"
            data:
              email: "Invalid email format"
              password: "Password must be at least 6 characters"
            timestamp: "2025-11-15T10:30:00Z"
            path: "/api/v1/register/buyer"
...